Added a Fragment for the Karnegram
Added a GridView to show images in

There is also a button to click that will take the user
to the camera.

When the user returns to the application from the camera the image
can be extracted and will be put in the gridview, just to show that
the image came from the camera and it works.